# Speedup Scan in different backends