… tests/scan/test_printing.py
… tests/tensor/test_subtensor.py